style: add dict.ts
